pyproject.toml
[project]
name = "emotional-speech-synthesis-pipeline"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alex Laletin", email = "alex.klark.laletin@yandex.com"},
]
readme = "README.md"
license = {text = "MIT"}

requires-python = "==3.11.*"
dependencies = [
    "click>=8.1.7",
    "pydub>=0.25.1",
    "requests>=2.32.3",
    "joblib>=1.4.2",
    "librosa>=0.10.2.post1",
    "tqdm>=4.67.0",
    "pandas>=2.2.3",
    "nvidia-pytriton>=0.5.13",
    "scipy>=1.14.1",
    "soundfile>=0.12.1",
    "jiwer>=3.0.5",
    "SQLAlchemy>=2.0.37",
    "python-dotenv>=1.0.1",
    "psycopg>=3.2.3",
]

[dependency-groups]
test = [
    "pytest>=8.3.3",
]
lint = [
    "ruff>=0.8.0",
]
dev = [
    "pre-commit>=4.0.1",
]

[tool.pdm]
distribution = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests",]

[tool.ruff]
line-length = 120